Tests/LibWeb/Text/expected/Worker/Worker-module.txt

Message from worker: Worker responding to: Hello :^)
